requirements.txt
discord.py==2.5.2
mysql-connector==2.2.9
python-dotenv==1.1.0
